new signup and login forms pages combined, full styling + toggle functions between forms and viewport adapatbility
