--- src/SignUp.css
@@ -1,224 +1,0 @@
-/* General Styles */
-html, body {
-  height: 100%;
-  margin: 0;
-  background-color: #F6F6F4;
-  font-family: 'Arial', sans-serif;
-  color: white;
-  display: flex;
-  flex-direction: column;
-}
-
-body {
-  display: flex;
-  flex-direction: column;
-  background-color: rgb(255, 255, 255) !important;
-  box-sizing: border-box;
-}
-
-main {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-}
-
-/* Styling the Signup-container */
-.signup-container {
-  padding: 25px;
-  width: 100%;
-  max-width: 400px;
-  height: auto;
-  border-radius: 10px;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-  text-align: center;
-  transition: transform 0.3s ease;
-  font-family: 'Arial', sans-serif;
-}
-
-.signup-container:hover {
-  transform: translateY(-3px);
-}
-
-.signup-container h1 {
-  font-size: 24px;
-  margin-bottom: 20px;
-  font-weight: normal; /* Ensure normal font weight */
-}
-
-.input-field {
-  margin-bottom: 15px;
-}
-
-.input-field label {
-  display: block;
-  margin-bottom: 8px;
-  font-weight: bold;
-  font-size: 16px;
-  color: white;
-  font-family: 'Arial', sans-serif;
-}
-
-.input-field input {
-  width: 100%;
-  padding: 12px;
-  margin-bottom: 20px;
-  border: 1px solid #ccc;
-  border-radius: 8px;
-  box-sizing: border-box;
-  font-size: 14px;
-  transition: border-color 0.3s ease;
-  background-color: white;
-  font-family: 'Arial', sans-serif;
-}
-
-.input-field input:focus {
-  border-color: #7c1816;
-  outline: none;
-}
-
-/* Button group styling */
-.button-group {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-}
-
-.signup-button {
-  background-color: #7c1816;
-  color: white;
-  border: none;
-  padding: 12px 20px;
-  border-radius: 20px;
-  cursor: pointer;
-  font-size: 16px;
-  transition: background-color 0.3s ease, transform 0.3s ease;
-  font-family: 'Arial', sans-serif;
-}
-
-.signup-button:hover {
-  background-color: #a02020;
-  box-shadow: 0 0 15px rgba(255, 50, 50, 1);
-}
-
-.admin-signup-link {
-  color: #7c1816;
-  text-decoration: none;
-  font-size: 16px;
-  transition: color 0.3s ease;
-  font-family: 'Arial', sans-serif;
-}
-
-.admin-signup-link:hover {
-  color: #5a1210;
-}
-
-/* Styling the Login-container */
-.login-container {
-  padding: 20px;
-  width: 100%;
-  max-width: 400px;
-  height: auto;
-  border-radius: 10px;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-  text-align: center;
-  transition: transform 0.3s ease;
-  font-family: 'Arial', sans-serif;
-}
-
-.login-container:hover {
-  transform: translateY(-5px);
-}
-
-.login-container h2 {
-  font-size: 20px;
-  margin-bottom: 10px;
-  font-weight: normal; /* Ensure normal font weight */
-}
-
-.login-container p {
-  font-size: 16px;
-  margin-bottom: 20px;
-  font-family: 'Arial', sans-serif;
-}
-
-.login-link {
-  background-color: #7c1816;
-  color: white;
-  padding: 12px 20px;
-  border-radius: 20px;
-  text-decoration: none;
-  font-size: 16px;
-  transition: background-color 0.3s ease, transform 0.3s ease;
-  box-shadow: 0 0 10px rgba(140, 0, 0, 0.8);
-  font-family: 'Arial', sans-serif;
-}
-
-.login-link:hover {
-  background-color: #a02020;
-  box-shadow: 0 0 15px rgba(255, 50, 50, 1);  
-}
-
-.signup-link {
-  background-color: #7c1816;
-  color: white;
-  padding: 12px 20px;
-  border-radius: 20px;
-  text-decoration: none;
-  font-size: 16px;
-  transition: background-color 0.3s ease, transform 0.3s ease;
-  box-shadow: 0 0 10px rgba(140, 0, 0, 0.8);
-  font-family: 'Arial', sans-serif;
-}
-
-/* Specific Popup Styles for AdminSignup */
-.admin-popup {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.admin-popup-content {
-  background: white;
-  padding: 20px;
-  border-radius: 10px;
-  text-align: center;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-  font-family: 'Arial', sans-serif;
-}
-
-.admin-popup-content h2 {
-  margin-bottom: 10px;
-  font-weight: normal; /* Ensure normal font weight */
-}
-
-.admin-popup-content p {
-  margin-bottom: 20px;
-  font-family: 'Arial', sans-serif;
-}
-
-.admin-popup-content button {
-  background-color: #7c1816;
-  color: white;
-  border: none;
-  padding: 10px 20px;
-  border-radius: 20px;
-  cursor: pointer;
-  font-size: 16px;
-  transition: background-color 0.3s ease, transform 0.3s ease;
-  font-family: 'Arial', sans-serif;
-}
-
-.admin-popup-content button:hover {
-  background-color: rgba(0, 0, 0, 0.1);
-  transform: translateY(-2px);
-}
-
