+++ src/VIP_lounge_landing.css
@@ -122,7 +122,7 @@
   background: #7a1917;
   color: #fff;
   border: none;
-  /* padding: 20px 30px; */
+  padding: 10px 20px;
   border-radius: 50px;
   cursor: pointer;
   font-weight: bold;
